add a syntax of custom_switch
