+++ Rantfile
@@ -53,15 +53,9 @@ import "directedrule"
 desc "Build trema."
 task :default => [
   :libtrema,
-#  :rubylib,
   :switch_manager,
   :switch,
   :packetin_filter,
-#  :tremashark,
-#  :packet_capture,
-#  :syslog_relay,
-#  :stdin_relay,
-#  :vendor,
   :examples
 ]
 
@@ -83,20 +77,16 @@ task :default => [
 #]
 # AFTER t-saito MOD
 task :examples => [
-#  "examples:cbench_switch",
   "examples:dumper",
   "examples:hello_trema",
   "examples:learning_switch",
   "examples:list_switches",
-#  "examples:multi_learning_switch",
   "examples:openflow_switch",
   "examples:openflow_message",
-#  "examples:packet_in",
   "examples:repeater_hub",
   "examples:packetin_filter_config",
   "examples:switch_info",
   "examples:switch_monitor",
-#  "examples:traffic_monitor",
 ]
 
 
@@ -237,7 +227,6 @@ import "sys/tgz"
 
 task :vendor => [
   "vendor:oflops",
-#  "vendor:openflow",
   "vendor:openvswitch",
   "vendor:phost",
 ]
@@ -245,7 +234,6 @@ task :vendor => [
 task :distclean => [
   "vendor:cmockery:distclean",
   "vendor:oflops:distclean",
-#  "vendor:openflow:distclean",
   "vendor:openvswitch:distclean",
   "vendor:phost:distclean",
 ]
@@ -711,11 +699,7 @@ end
 ################################################################################
 
 openflow_messages = [
-#  "echo_reply",
-#  "echo_request",
-#  "features_request",
   "hello",
-#  "set_config",
 ]
 
 openflow_message_source_dir = "src/examples/openflow_message"
