+++ ruby/trema/dsl/syntax.rb
@@ -27,6 +27,7 @@ require "trema/dsl/run"
 require "trema/dsl/switch"
 require "trema/dsl/vhost"
 require "trema/dsl/vswitch"
+require "trema/dsl/custom-switch"
 require "trema/hardware-switch"
 require "trema/host"
 require "trema/link"
@@ -35,7 +36,7 @@ require "trema/open-vswitch"
 require "trema/packetin-filter"
 require "trema/ruby-switch"
 require "trema/switch-manager"
-require "trema/test-switch"
+require "trema/custom-switch"
 
 
 module Trema
@@ -67,12 +68,7 @@ module Trema
       def vswitch name = nil, &block
         stanza = Trema::DSL::Vswitch.new( name )
         stanza.instance_eval( &block )
-
-        if stanza[ :stub ].nil?
-          Trema::OpenVswitch.new stanza, @config.port
-        else
-          Trema::TestSwitch.new stanza
-        end
+        Trema::OpenVswitch.new stanza, @config.port
       end
 
 
@@ -80,6 +76,13 @@ module Trema
         stanza = Trema::DSL::Rswitch.new( name )
         stanza.instance_eval( &block )
         Trema::RubySwitch.new( stanza )
+      end
+
+
+      def custom_switch name = nil, &block
+        stanza = Trema::DSL::CustomSwitch.new( name )
+        stanza.instance_eval( &block )
+        Trema::CustomSwitch.new stanza
       end
