+++ src/examples/openflow_switch/hello_switch.c
@@ -1,7 +1,7 @@
 /*
  * A simple switch that has minimum function to test hello message.
  *
- * Copyright (C) 2008-2012 NEC Corporation
+ * Copyright (C) 2012 Hiroyasu OHYAMA
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License, version 2, as
